+++ tasty-plutus/src/Test/Tasty/Plutus/Script/Unit.hs
@@ -2,7 +2,7 @@
 
 {- |
  Module: Test.Tasty.Plutus.Script.Unit
- Copyright: (C) MLabs 2021
+ Copyright: (C) MLabs 2021-2022
  License: Apache 2.0
  Maintainer: Koz Ross <koz@mlabs.city>
  Portability: GHC only
@@ -40,9 +40,22 @@ import Data.Tagged (Tagged (Tagged))
 import Data.Text (Text)
 import Data.Vector (Vector)
 import Data.Vector qualified as Vector
-import Plutus.V1.Ledger.Api (ScriptContext)
+import Plutus.V1.Ledger.Api (
+  ExBudget (ExBudget),
+  ExCPU (ExCPU),
+  ExMemory (ExMemory),
+  ScriptContext,
+ )
+import Plutus.V1.Ledger.Scripts (
+  ScriptError (
+    EvaluationError,
+    EvaluationException,
+    MalformedScript
+  ),
+ )
 import Test.Plutus.ContextBuilder (
   ContextBuilder,
+  Naming,
   Purpose (ForMinting, ForSpending),
   TransactionConfig,
  )
@@ -58,12 +71,16 @@ import Test.Tasty.Plutus.Internal.Env (
   getScriptResult,
   prepareConf,
  )
+import Test.Tasty.Plutus.Internal.Estimate (minterEstimate, spenderEstimate)
 import Test.Tasty.Plutus.Internal.Feedback (
   didn'tLog,
   doPass,
   dumpState,
+  errorNoEstimate,
+  explainFailureEstimation,
   malformedScript,
   noParse,
+  reportBudgets,
   scriptException,
   unexpectedFailure,
   unexpectedSuccess,
@@ -86,6 +103,7 @@ import Test.Tasty.Plutus.Internal.WithScript (
  )
 import Test.Tasty.Plutus.Options (
   Fee,
+  PlutusEstimate (EstimateOnly, NoEstimates),
   PlutusTracing,
   ScriptInputPosition,
   TestCurrencySymbol,
@@ -96,6 +114,7 @@ import Test.Tasty.Plutus.Options (
 import Test.Tasty.Plutus.TestData (
   Outcome (Fail, Pass),
   TestData (MintingTest, SpendingTest),
+  TestItems (ItemsForMinting, ItemsForSpending),
  )
 import Test.Tasty.Providers (
   IsTest (run, testOptions),
@@ -110,14 +129,14 @@ import Type.Reflection (Typeable)
 {- | Specify that, given this test data and context, the validation should
  succeed.
 
- @since 3.0
+ @since 9.0
 -}
 shouldValidate ::
-  forall (p :: Purpose).
-  (Typeable p) =>
+  forall (p :: Purpose) (n :: Naming).
+  (Typeable p, Typeable n) =>
   String ->
   TestData p ->
-  ContextBuilder p ->
+  ContextBuilder p n ->
   WithScript p ()
 shouldValidate = addUnitTest Pass Nothing
 
@@ -127,28 +146,28 @@ shouldValidate = addUnitTest Pass Nothing
  * The validation should succeed; and
  * The trace that results should satisfy the predicate.
 
- @since 3.3
+ @since 9.0
 -}
 shouldValidateTracing ::
-  forall (p :: Purpose).
-  (Typeable p) =>
+  forall (p :: Purpose) (n :: Naming).
+  (Typeable p, Typeable n) =>
   String ->
   (Vector Text -> Bool) ->
   TestData p ->
-  ContextBuilder p ->
+  ContextBuilder p n ->
   WithScript p ()
 shouldValidateTracing name f = addUnitTest Pass (Just f) name
 
 {- | Specify that, given this test data and context, the validation should fail.
 
- @since 3.0
+ @since 9.0
 -}
 shouldn'tValidate ::
-  forall (p :: Purpose).
-  (Typeable p) =>
+  forall (p :: Purpose) (n :: Naming).
+  (Typeable p, Typeable n) =>
   String ->
   TestData p ->
-  ContextBuilder p ->
+  ContextBuilder p n ->
   WithScript p ()
 shouldn'tValidate = addUnitTest Fail Nothing
 
@@ -158,28 +177,28 @@ shouldn'tValidate = addUnitTest Fail Nothing
  * The validation should fail; and
  * The resulting trace should satisfy the predicate.
 
- @since 3.3
+ @since 9.0
 -}
 shouldn'tValidateTracing ::
-  forall (p :: Purpose).
-  (Typeable p) =>
+  forall (p :: Purpose) (n :: Naming).
+  (Typeable p, Typeable n) =>
   String ->
   (Vector Text -> Bool) ->
   TestData p ->
-  ContextBuilder p ->
+  ContextBuilder p n ->
   WithScript p ()
 shouldn'tValidateTracing name f = addUnitTest Fail (Just f) name
 
 -- Helpers
 
 addUnitTest ::
-  forall (p :: Purpose).
-  (Typeable p) =>
+  forall (p :: Purpose) (n :: Naming).
+  (Typeable p, Typeable n) =>
   Outcome ->
   Maybe (Vector Text -> Bool) ->
   String ->
   TestData p ->
-  ContextBuilder p ->
+  ContextBuilder p n ->
   WithScript p ()
 addUnitTest out trace name td cb = case td of
   SpendingTest {} -> WithSpending $ do
@@ -189,53 +208,61 @@ addUnitTest out trace name td cb = case td of
     tt <- asks (singleTest name . Minter out trace td cb)
     tell . Seq.singleton $ tt
 
-data ScriptTest (p :: Purpose) where
+data ScriptTest (p :: Purpose) (n :: Naming) where
   Spender ::
-    forall (d :: Type) (r :: Type).
+    forall (d :: Type) (r :: Type) (n :: Naming).
     Outcome ->
     Maybe (Vector Text -> Bool) ->
     TestData ( 'ForSpending d r) ->
-    ContextBuilder ( 'ForSpending d r) ->
+    ContextBuilder ( 'ForSpending d r) n ->
     TestScript ( 'ForSpending d r) ->
-    ScriptTest ( 'ForSpending d r)
+    ScriptTest ( 'ForSpending d r) n
   Minter ::
-    forall (r :: Type).
+    forall (r :: Type) (n :: Naming).
     Outcome ->
     Maybe (Vector Text -> Bool) ->
     TestData ( 'ForMinting r) ->
-    ContextBuilder ( 'ForMinting r) ->
+    ContextBuilder ( 'ForMinting r) n ->
     TestScript ( 'ForMinting r) ->
-    ScriptTest ( 'ForMinting r)
+    ScriptTest ( 'ForMinting r) n
 
-data UnitEnv (p :: Purpose) = UnitEnv
+getOutcome ::
+  forall (p :: Purpose) (n :: Naming).
+  ScriptTest p n ->
+  Outcome
+getOutcome = \case
+  Spender out _ _ _ _ -> out
+  Minter out _ _ _ _ -> out
+
+data UnitEnv (p :: Purpose) (n :: Naming) = UnitEnv
   { envOpts :: OptionSet
-  , envScriptTest :: ScriptTest p
+  , envScriptTest :: ScriptTest p n
   }
 
 getShouldChat ::
-  forall (p :: Purpose).
-  UnitEnv p ->
+  forall (p :: Purpose) (n :: Naming).
+  UnitEnv p n ->
   PlutusTracing
 getShouldChat = lookupOption . envOpts
 
 getConf ::
-  forall (p :: Purpose).
-  UnitEnv p ->
+  forall (p :: Purpose) (n :: Naming).
+  UnitEnv p n ->
   TransactionConfig
 getConf = prepareConf envOpts
 
 getCB ::
-  forall (p :: Purpose).
-  UnitEnv p ->
-  ContextBuilder p
+  forall (p :: Purpose) (n :: Naming).
+  UnitEnv p n ->
+  ContextBuilder p n
 getCB =
   envScriptTest >>> \case
     Spender _ _ _ cb _ -> cb
     Minter _ _ _ cb _ -> cb
 
 getTestData ::
-  forall (p :: Purpose).
-  UnitEnv p ->
+  forall (p :: Purpose) (n :: Naming).
+  UnitEnv p n ->
   TestData p
 getTestData =
   envScriptTest >>> \case
@@ -243,8 +270,8 @@ getTestData =
     Minter _ _ td _ _ -> td
 
 getScript ::
-  forall (p :: Purpose).
-  UnitEnv p ->
+  forall (p :: Purpose) (n :: Naming).
+  UnitEnv p n ->
   SomeScript p
 getScript =
   envScriptTest >>> \case
@@ -252,8 +279,8 @@ getScript =
     Minter _ _ _ _ mp -> SomeMinter . getTestMintingPolicy $ mp
 
 getMPred ::
-  forall (p :: Purpose).
-  UnitEnv p ->
+  forall (p :: Purpose) (n :: Naming).
+  UnitEnv p n ->
   Maybe (Vector Text -> Bool)
 getMPred =
   envScriptTest >>> \case
@@ -261,8 +288,8 @@ getMPred =
     Minter _ mPred _ _ _ -> mPred
 
 getExpected ::
-  forall (p :: Purpose).
-  UnitEnv p ->
+  forall (p :: Purpose) (n :: Naming).
+  UnitEnv p n ->
   Outcome
 getExpected =
   envScriptTest >>> \case
@@ -270,31 +297,51 @@ getExpected =
     Minter expected _ _ _ _ -> expected
 
 getSC ::
-  forall (p :: Purpose).
-  UnitEnv p ->
+  forall (p :: Purpose) (n :: Naming).
+  UnitEnv p n ->
   ScriptContext
 getSC = getScriptContext getConf getCB getTestData
 
 getDumpedState ::
-  forall (p :: Purpose).
+  forall (p :: Purpose) (n :: Naming).
   [Text] ->
-  UnitEnv p ->
+  UnitEnv p n ->
   Doc
 getDumpedState = dumpState getConf getCB getTestData
 
-instance (Typeable p) => IsTest (ScriptTest p) where
-  run opts vt _ = pure $ case result of
-    Left err -> runReader (handleError err) env
-    Right logs -> runReader (deliverResult logs) env
+instance (Typeable p, Typeable n) => IsTest (ScriptTest p n) where
+  run opts vt _ = pure $ case lookupOption opts of
+    EstimateOnly -> case getOutcome vt of
+      Fail -> testPassed explainFailureEstimation
+      _ -> case tryEstimate of
+        Left err -> testFailed $ case err of
+          EvaluationError logs msg -> errorNoEstimate logs msg
+          EvaluationException name msg -> scriptException name msg
+          MalformedScript msg -> malformedScript msg
+        Right (ExBudget (ExCPU bCPU) (ExMemory bMem)) ->
+          testPassed $ reportBudgets (fromIntegral bCPU) (fromIntegral bMem)
+    NoEstimates -> runReader go env
     where
-      env :: UnitEnv p
+      tryEstimate :: Either ScriptError ExBudget
+      tryEstimate = case vt of
+        Spender out _ td cb ts -> case td of
+          SpendingTest dat red v ->
+            let ti = ItemsForSpending dat red v cb out
+             in spenderEstimate opts ts ti
+        Minter out _ td cb ts -> case td of
+          MintingTest red tasks ->
+            let ti = ItemsForMinting red tasks cb out
+             in minterEstimate opts ts ti
+      go :: Reader (UnitEnv p n) Result
+      go = case getScriptResult getScript getTestData (getContext getSC) env of
+        Left err -> handleError err
+        Right logs -> deliverResult logs
+      env :: UnitEnv p n
       env =
         UnitEnv
           { envOpts = opts
           , envScriptTest = vt
           }
-      result :: Either ScriptResult [Text]
-      result = getScriptResult getScript getTestData (getContext getSC) env
   testOptions =
     Tagged
       [ Option @Fee Proxy
@@ -304,12 +351,13 @@ instance (Typeable p) => IsTest (ScriptTest p) where
       , Option @TestValidatorHash Proxy
       , Option @PlutusTracing Proxy
       , Option @ScriptInputPosition Proxy
+      , Option @PlutusEstimate Proxy
       ]
 
 handleError ::
-  forall (p :: Purpose).
+  forall (p :: Purpose) (n :: Naming).
   ScriptResult ->
-  Reader (UnitEnv p) Result
+  Reader (UnitEnv p n) Result
 handleError = \case
   ScriptFailed logs msg ->
     asks getExpected >>= \case
@@ -323,16 +371,20 @@ handleError = \case
     pure . testFailed $ malformedScript msg
 
 deliverResult ::
-  forall (p :: Purpose).
+  forall (p :: Purpose) (n :: Naming).
   [Text] ->
-  Reader (UnitEnv p) Result
+  Reader (UnitEnv p n) Result
 deliverResult logs = do
   expected <- asks getExpected
   case expected of
     Fail -> asks (testFailed . unexpectedSuccess (getDumpedState logs))
     Pass -> asks getMPred >>= (`tryPass` logs)
 
-tryPass :: Maybe (Vector Text -> Bool) -> [Text] -> Reader (UnitEnv p) Result
+tryPass ::
+  forall (p :: Purpose) (n :: Naming).
+  Maybe (Vector Text -> Bool) ->
+  [Text] ->
+  Reader (UnitEnv p n) Result
 tryPass mPred logs = case mPred of
   Nothing -> asks (testPassed . doPass getShouldChat logs)
   Just f ->