+++ tasty-plutus/CHANGELOG.md
@@ -4,12 +4,28 @@ This format is based on [Keep A Changelog](https://keepachangelog.com/en/1.0.0).
 
 ## Unreleased
 
-## 8.1 - 2022-03-20
+## 9.1 - 2022-03-22
 
 ### Added
 
 * `mkTestValidatorUntyped` and `mkTestMintingPolicyUntyped` for constructing
   `TestScript`s out of untyped code
+
+## 9.0 -- 2022-03-22
+
+### Changed
+
+* Support for named and anonymous `ContextBuilder`s, in line with changes in
+  `plutus-context-builder`. These changes are technically breaking, but
+  shouldn't have issues in themselves, as long as the `ContextBuilder`s are
+  updated.
+
+## 8.1 - 2022-03-16
+
+### Added
+
+* `PlutusEstimate` option, which displays estimated memory and CPU use for
+  scripts instead of testing them.
 
 ## 8.0 - 2022-02-08